refactor(partyroom): be-a-host 핸들러를 useBeAHost hook 으로 추출 (동작 0)
